+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v5.5.0
 **ADRSolver**
 - Add support for spatially-constant, but variable direction, diffusion to
   ADRSolver. (!1669)
+- Inline Vmath library (!1667)
 
 **CardiacEPSolver**
 - Fix cell model history point filter output after base class change (!1342)